requirements.txt
django=3.0.3
djangorestframework=3.11